:bug: Fix issue #135 - return correct null value.
The AbstractObjectAttributeDescription.getString() method returns
`String.valueOf((T) data[id])` even if `data[id]` is null. This returns
the String "null" when a float_or_null value is null, which in turn
causes Float.parseFloat() at line 90 in RestUtilities.addData() to
throw a java.lang.NumberFormatException: For input string: "null".

This commit checks for a null value and returns either a non-null
value or null.