:bug: Correctly return error messages to REST clients.
The Python requests client expects the first line of an HTTP response
message to be of the form

Status-Line = HTTP-Version SP Status-Code SP Reason-Phrase CRLF

as per RFC 2616 section 6.1. However, there is nothing in the servlet
code to properly set the reason phrase. This commit catches exceptions
in the called code and correctly sets the response using
response.sendError().

However, the call to response.SendError sort of works: it sets the
status and status phrase, but then errors with
"java.lang.NoSuchMethodError: javax.servlet.http.HttpServletResponse.getStatus()I".
This indicates mismatched jars: see for example:
https://stackoverflow.com/questions/44730096/javax-servlet-http-httpservletresponse-getstatus-not-found .
Some work-around code catches the exception and continues.
When the jars are sorted out, the extra code can be removed.