:green_heart: Stale detection action workflow
Add a stale detection action workflow that will run once a day. With the growing number of issue and PRs we need to keep on top of them and make sure we don't let them hang around forever or be forgotten about.

This is a very lenient numbers and hopefully we don't see this too often!